chore: remove docgen (#14)
It's not a good contributor experience to send us a spelling fix in a starlark comment and get a red PR because the equivalent Markdown isn't updated. There's not a nice way to 'accept fix' in this case.
Stardoc is just too complicated to setup and maintain in every ruleset repo. I'm making a centralized starlark docgen pipeline for this in https://github.com/alexeagle/doc.bzl